Bootstrap Compatibility
The editor is no longer bound to the bottom of the page, but can be
added to any div on the page.  Dynamic sizing is demonstated through
button actions, as well as the ability to clear localStorage.

It also has a json viewer to see all the settings used to create the
timeline, which could then be saved to a database.